WIP: remove the original module
